# src/swagger-config.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "CASMI APIs",
        "description": "CASMI: Cloud Agnostic Storage Management Interface",
        "version": "1.0"
    },
    "servers": [{ "url": "http://localhost:8000/api/v1" }],
    "paths": {
        "/files": {
            "get": {
                "summary": "Summary",
                "description": "Description",
                "tags": ["Files"],
                "security": [{ "Bearer token": [] }],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "files": {
                                            "type": "array",
                                            "items": {
                                                "type": "object",
                                                "properties": {
                                                    "id": { "type": "integer" },
                                                    "name": { "type": "string" },
                                                    "size": { "type": "integer" },
                                                    "provider_key": { "type": "string" },
                                                    "folder_id": { "type": "integer" },
                                                    "store_id": { "type": "integer" },
                                                    "space_id": { "type": "integer" },
                                                    "space": { "type": "string" },
                                                    "provider_id": { "type": "integer" },
                                                    "provider": { "type": "string" }
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/files/{id}": {
            "get": {
                "summary": "Summary",
                "description": "Description",
                "tags": ["Files"],
                "security": [{ "Bearer token": [] }],
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "description": "The file ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "file": {
                                            "type": "object",
                                            "properties": {
                                                "id": { "type": "integer" },
                                                "name": { "type": "string" },
                                                "size": { "type": "integer" },
                                                "provider_key": { "type": "string" },
                                                "folder_id": { "type": "integer" },
                                                "store_id": { "type": "integer" },
                                                "space_id": { "type": "integer" },
                                                "space": { "type": "string" },
                                                "provider_id": { "type": "integer" },
                                                "provider": { "type": "string" }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Client error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": { "type": "boolean" },
                                        "message": { "type": "string" }
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "Bearer token": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}